% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionorder.R
\name{buttonkrmesproductionordersrcbillUI_left}
\alias{buttonkrmesproductionordersrcbillUI_left}
\title{操作区域}
\usage{
buttonkrmesproductionordersrcbillUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
