% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionorder.R
\name{buttonkrmesproductionordersrcbillUI_right}
\alias{buttonkrmesproductionordersrcbillUI_right}
\title{预览区域}
\usage{
buttonkrmesproductionordersrcbillUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
