% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionreturn.R
\name{buttonkrmesproductionreturnsrcbillUI_left}
\alias{buttonkrmesproductionreturnsrcbillUI_left}
\title{操作区域}
\usage{
buttonkrmesproductionreturnsrcbillUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
