% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionreturn.R
\name{buttonkrmesproductionreturnsrcbillUI_right}
\alias{buttonkrmesproductionreturnsrcbillUI_right}
\title{预览区域}
\usage{
buttonkrmesproductionreturnsrcbillUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
